requirements.txt
Django==3.2.3
psycopg2==2.8.6
python-decouple==3.4
django-heroku==0.3.1
gunicorn==20.1.0
django-inlinecss==0.3.0
django-allauth==0.44.0